+++ BE/httc-sport/src/main/java/com/fpoly/httc_sport/repository/RentInfoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fpoly.httc_sport.entity.RentInfo;
 
-public interface RentInfoRepository extends JpaRepository<RentInfo, Long> {
+public interface RentInfoRepository extends JpaRepository<RentInfo, Integer> {
 	List<RentInfo> findByPitchId(Integer pitchId);
 	List<RentInfo> findByUserId(String userId);
 	Boolean existsByPitchIdEqualsAndEmailEqualsAndRentedAtEqualsAndStartTimeLessThanEqualAndEndTimeGreaterThanEqual(Integer pitchId, String email, LocalDate date, LocalTime time, LocalTime time2);
